back-end/src/requirements.txt
fastapi
sqlalchemy 
python-dotenv 
psycopg2-binary
gunicorn 
uvicorn
python-multipart 
pydantic[all] 
pydantic==1.10.8
passlib 
bcrypt
pydub==0.25.1
numpy
requests
pkgconfig=1.5.5
Pillow
tensorflow
opencv-python
scikit-learn
matplotlib
pandas
rembg